auth flow

+++ database/migrations/2023_09_29_031434_create_scores_table.php
@@ -13,13 +13,13 @@ return new class extends Migration
     {
         Schema::create('scores', function (Blueprint $table) {
             $table->id();
-            $table->integer('mark');
-            $table->integer('type');
+            $table->double('mark');
+            $table->tinyInteger('type');
             $table->integer('exams_id');
             $table->integer('students_id');
             $table->integer('subjects_id');
             $table->timestamps();
-            $table->softdeletes();     
+            $table->softdeletes();
         });
     }
